Redid some formatting and solved some warnings of the more pedantic type

+++ lib/my/src/my_nbr_to_string_base_width.c
@@ -29,7 +29,7 @@ static char *finish_conversion(const char *converted_number_ptr,
     const char *buffer_end, size_t width)
 {
     char *result;
-    size_t converted_number_size = (buffer_end - converted_number_ptr);
+    size_t converted_number_size = (size_t)(buffer_end - converted_number_ptr);
     if (converted_number_size > width)
         width = converted_number_size;
